Replaced Vector3D with M$ implementation

+++ Yupi.Model/Test/ValidateModel.cs
@@ -10,13 +10,13 @@ namespace Yupi.Model
 		{
 			foreach (RoomModel model in RoomModel.GetAll()) {
 				string[] rows = model.Heightmap.Split ('\r');
-				if (rows [model.Door.Y] [model.Door.X] == 'x') {
+				if (rows [(int)model.Door.Y] [(int)model.Door.X] == 'x') {
 
 					Console.WriteLine (model.DisplayName);
 					PrintFix (rows, model);
 
-				} else if (rows [model.Door.Y] [model.Door.X] != rows [model.Door.Y] [model.Door.X + 1]) {
-					Console.WriteLine ("Invalid door height: " + model.DisplayName + " should be [ " + rows [model.Door.Y] [model.Door.X + 1] + "]");
+				} else if (rows [(int)model.Door.Y] [(int)model.Door.X] != rows [(int)model.Door.Y] [(int)model.Door.X + 1]) {
+					Console.WriteLine ("Invalid door height: " + model.DisplayName + " should be [ " + rows [(int)model.Door.Y] [(int)model.Door.X + 1] + "]");
 					PrintFix (rows, model);
 				}
 			}
@@ -24,9 +24,9 @@ namespace Yupi.Model
 
 		private static void PrintFix (string[] rows, RoomModel model)
 		{
-			char[] doorRow = rows [model.Door.Y].ToCharArray ();
+			char[] doorRow = rows [(int)model.Door.Y].ToCharArray ();
 
-			doorRow [model.Door.X] = rows [model.Door.Y] [model.Door.X + 1];
+			doorRow [(int)model.Door.X] = rows [(int)model.Door.Y] [(int)model.Door.X + 1];
 
 			for (int i = 0; i < rows.Length; i++) {
 				if (i == model.Door.X) {
